cell_transmission_model/pyproject.toml
[project]
name = "cell-transmission-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jean Doig", email = "jcdoig@berkeley.edu" }
]
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"
